Finish wiring up app
