+++ apps/app/app/(unauthenticated)/sign-up/[[...sign-up]]/page.tsx
@@ -9,7 +9,7 @@ const description = 'Sign Up to your account.';
 export const metadata: Metadata = createMetadata({ title, description });
 
 const SignUpPage: FC = () => (
-  <div className="flex min-h-[calc(100dvh-61px)] w-full flex-col items-center justify-center">
+  <div className="flex min-h-dvh w-full flex-col items-center justify-center bg-secondary">
     <SignUp />
   </div>
 );
